% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_met.R
\name{bayes_met}
\alias{bayes_met}
\title{Bayesian model for multi-environment trials}
\usage{
bayes_met(
  data,
  gen,
  env,
  rept,
  trait,
  hyperparam = "default",
  sigma.dist = c("cauchy", "cauchy"),
  mu.dist = c("normal", "cauchy"),
  gli.dist = c("normal", "cauchy"),
  reg = NULL,
  iter = 4000,
  cores = 4,
  chain = 4,
  ...
)
}
\arguments{
\item{data}{A dataframe containing the observations}

\item{gen}{A string vector. The first element is the name of the
column that corresponds to the evaluated genotype. The second element is the
prior probability distribution of this parameter. The third element is the
hyperprior of this parameter. This order must be followed.}

\item{env}{A string vector. The first element is the name of the
column that corresponds to the environments. The second element is the
prior probability distribution of this parameter. The third element is the
hyperprior of this parameter. This order must be followed.}

\item{rept}{A vector, a list or NULL. If the trial is randomized in complete blocks,
rept will be a vector with the first element representing the name of the column
that corresponds to the blocks, the second element being the prior probability
distribution of the block parameter, and the third element representing the
hyperprior of this parameter. If the trial is randomized in a lattice,
rept will be a list containing two vectors with the same structure previously
mentioned: a vector for replicate effects, and a vector for block effects.
If the data do not have replicates, rept will be NULL.}

\item{trait}{A character representing the name of the column that
corresponds to the analysed variable.}

\item{hyperparam}{A numeric representing the global hyperparameter. If no number
is give (so hyperparam = 'default'), the function will use max(trait) * 10
as the global hyperparameter.}

\item{sigma.dist}{A string vector containing the prior probability distribution
(1st element) and the hyperprior (2nd element) of the residual effects.
The default is c("cauchy","cauchy").}

\item{mu.dist}{A string vector containing the prior probability distribution
(1st element) and the hyperprior (2nd element) of the intercept. The default
is c("normal","cauchy").}

\item{gli.dist}{A string vector containing the prior probability distribution
(1st element) and the hyperprior (2nd element) of the genotype-by-environment
interaction. The default is c("normal","cauchy").}

\item{reg}{A list containing two string vectors. The first vector have the name
of the column that contain the information of region in the first position, the
prior probability of this parameter in the second position, and its hyperprior
in the third position. The second vector have the the prior probability and
hyperprior of the genotype-by-region interaction effects, in the first and
second position, respectively. If there are no information about regions in
the dataset, reg = NULL (default).}

\item{iter}{Inherited from the "sampling" function of the "rstan" package.
A positive integer specifying the number of iterations for each chain
(including warmup). The default is 2000.}

\item{cores}{Inherited from the "sampling" function of the "rstan" package.
A positive integer specifying the number of cores to use when executing the
chains in parallel (defaults to 1).}

\item{chain}{Inherited from the "sampling" function of the "rstan" package.
A positive integer specifying the number of Markov chains. The default is 4.}

\item{...}{Additional arguments passed to the "sampling" function.
For more information, see "sampling" manual}
}
\value{
The function returns an object of S4 class stanfit containing the
fitted results
}
\description{
This function runs a Bayesian model for analysing data from Multi-environment
trials using RStan, the R interface to Stan.
}
\examples{
\dontrun{
mod = bayes_met(data = df, gen = c("Hybrid", "normal", "cauchy"),
                env = c("Location", "normal", "cauchy"),
                rept = list(c("Rep", "normal", "cauchy"), c("Block", "normal", "cauchy")),
                trait = "GY", hyperparam = "default", sigma.dist = c("cauchy", "cauchy"),
                mu.dist = c("normal", "normal"), gli.dist = c("normal", "cauchy"),
                reg = list(c("Region", "normal", "cauchy"), c("normal", "cauchy")),
                iter = 8000, cores = 4, chain = 4)
                }
}
