% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_outs.R
\name{extr_outs}
\alias{extr_outs}
\title{Extracting outputs from \code{\link[=bayes_met]{bayes_met()}} objects}
\usage{
extr_outs(
  data,
  trait,
  model,
  probs = c(0.025, 0.975),
  plots = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{data}{A data frame containing the observations}

\item{trait}{A character representing the name of the column that
corresponds to the analysed trait}

\item{model}{An object of class \code{stanfit}, obtained using \code{\link[=bayes_met]{bayes_met()}}}

\item{probs}{A vector with two elements representing the probabilities
(in decimal scale) that will be considered for computing the quantiles.}

\item{verbose}{A logical value. If \code{TRUE}, the function will indicate the
completed steps. Defaults to \code{FALSE}}

\item{...}{Passed to \code{\link[rstan:stan_plot_diagnostics]{rstan::stan_diag()}}}

\item{plot}{Should the function build diagnostic plots? If \code{TRUE} (default),
it will provide histograms, density plots and traceplots of each effect declared in the model.}
}
\value{
The function returns an object of class \code{extr}, which is a list with:
\itemize{
\item \code{post} : a list with the posterior of the effects, and the data
generated by the model
\item \code{map} : a list with the maximum posterior values of each effect
\item \code{ppcheck} : a matrix containing the p-values of maximum, minimum,
median, mean and standard deviation; effective number of parameters, WAIC2
value, Rhat and effective sample size.
\item \code{plots} : a list with three types of ggplots: histograms, trace plots and density
plots. These will be available for all effects declared at the \code{effects} argument.
}
}
\description{
This function extracts outputs of the Bayesian model fitted
using \code{\link[=bayes_met]{bayes_met()}}, and provides some diagnostics.
}
\details{
More details about the usage of \code{extr_outs} and other functions of
the \code{ProbBreed} package can be found at \url{https://saulo-chaves.github.io/ProbBreed_site/}.
}
\examples{
\donttest{
mod = bayes_met(data = soy,
                gen = "Gen",
                loc = "Loc",
                repl = NULL,
                year = NULL,
                reg = NULL,
                res.het = FALSE,
                trait = 'Y',
                iter = 6000, cores = 4, chains = 4)

outs = extr_outs(data = soy, trait = "Y", model = mod,
                 probs = c(0.05, 0.95), plots = TRUE,
                 verbose = TRUE)
                 }
}
\seealso{
\code{\link[rstan:stan_plot_diagnostics]{rstan::stan_diag()}}, \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, \code{\link[rstan:check_hmc_diagnostics]{rstan::check_hmc_diagnostics()}}
}
