% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.R
\docType{data}
\name{maize}
\alias{maize}
\title{Maize real dataset}
\format{
\subsection{\code{maize}}{

A data frame with 823 rows and 6 columns:
\describe{
\item{Location}{16 locations}
\item{Region}{5 regions}
\item{Rep}{2 replicates}
\item{Block}{6 blocks}
\item{Hybrid}{36 genotypes}
\item{GY}{Grain yield (phenotypes)}
}
}
}
\source{
\describe{
Dias, K. O. G, Santos, J. P. R., Krause, M. D., Piepho, H. -P., Guimarães, L. J. M.,
Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts
for cultivar recommendation in multi-environment trials. \emph{Theoretical and
Applied Genetics}, 133(2):443-455. \doi{10.1007/s00122-022-04041-y}
}
}
\usage{
maize
}
\description{
This dataset belongs to value of cultivation and use maize trials of
Embrapa Maize and Sorghum, and was used by Dias et al. (2022).
It contains the grain yield of 32 single-cross hybrids and four commercial checks
(36 genotypes in total) evaluated in 16 locations across
five regions or mega-environments. These trials were laid out in incomplete blocks
design, using a block size of 6 and two replications per trial.
}
\keyword{datasets}
