% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_outs.R
\name{print.extr}
\alias{print.extr}
\title{Print an object of class \code{extr}}
\usage{
\method{print}{extr}(obj)
}
\arguments{
\item{obj}{An object of class \code{extr}}
}
\description{
Print a \code{extr} object in R console
}
\examples{
\donttest{
mod = bayes_met(data = maize,
                gen = "Hybrid",
                loc = "Location",
                repl = c("Rep", "Block"),
                year = NULL,
                reg = 'Region',
                res.het = FALSE,
                trait = 'GY',
                iter = 6000, cores = 4, chains = 4)

outs = extr_outs(data = maize, trait = "GY", model = mod,
                 probs = c(0.05, 0.95), plots = TRUE,
                 verbose = TRUE)
print(outs)
                 }
}
\seealso{
\link{extr_out}
}
