% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_sup.R
\name{print.probsup}
\alias{print.probsup}
\title{Print an object of class \code{probsup}}
\usage{
\method{print}{probsup}(x, ...)
}
\arguments{
\item{x}{An object of class \code{probsup}}

\item{...}{currently not used}
}
\description{
Print a \code{probsup} object in R console
}
\seealso{
\link{prob_sup}
}
