% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbBreed-package.R
\docType{package}
\name{ProbBreed-package}
\alias{ProbBreed}
\alias{ProbBreed-package}
\title{ProbBreed: Probability of Superior Performance and Stability of Selection Candidates Evaluated at Multi-Environment Trials}
\description{
Compute the probability of superior performance and superior stability of selection candidates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saulo-chaves/ProbBreed}
}

}
\author{
\strong{Maintainer}: Saulo Chaves \email{saulo.chaves@ufv.br} (\href{https://orcid.org/0000-0002-0694-1798}{ORCID})

Authors:
\itemize{
  \item Kaio Dias \email{kaio.o.dias@ufv.br} (\href{https://orcid.org/0000-0002-9171-1021}{ORCID}) [copyright holder]
  \item Matheus Krause \email{mdkrause@iastate.edu} (\href{https://orcid.org/0000-0002-9171-1021}{ORCID})
}

}
\keyword{internal}
