% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soy_sim.R
\docType{data}
\name{soy_sim}
\alias{soy_sim}
\title{Soybean simulated data set}
\format{
\subsection{\code{soy_sim}}{

A data frame with 1512 rown and 4 columns:
\describe{
\item{ID}{62 evaluated genotypes}
\item{pheno}{1512 phenotypes}
\item{L}{8 environments}
\item{rep}{3 replicates}
}
}
}
\source{
\describe{
online soon
}
}
\usage{
soy_sim
}
\description{
This data set contains 1,512 simulated phenotypes of 62 soybean genotypes
evaluated 8 environments, at trials designed in randomized complete blocks
design with 3 replicates.
}
\keyword{datasets}
