% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_prob.R
\name{cond_prob}
\alias{cond_prob}
\title{Conditional probabilities of superior performance}
\usage{
cond_prob(
  data,
  trait,
  gen,
  env,
  reg = NULL,
  extr_outs,
  int = 0.2,
  save.df = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the observations}

\item{trait}{A character representing the name of the column that
corresponds to the analysed variable}

\item{gen}{A character representing the name of the column that
corresponds to the evaluated genotypes}

\item{env}{A character representing the name of the column that
corresponds to the environments}

\item{reg}{A character representing the name of the column that
corresponds to the regions. NULL otherwise (default)}

\item{extr_outs}{An object from the 'extr_outs' function}

\item{int}{numeric A number representing the selection intensity
(superior limit = 1)}

\item{save.df}{A logical value indicating if the data frames with the marginal
probability of each genotype and the pairwise probabilities should be saved at
the work directory.}

\item{interactive}{A logical value indicating if the plots should be interactive.
If TRUE, the function loads the "plotly" package and uses the "ggplotly" command.}
}
\value{
The function returns a list with:
\itemize{
\item \code{risk.plot} a plot representing the Bayesian Eskridge's Risk
index (Adapted from Eskridge (1990))
\item \code{psp_env_reg.mat} : a matrix containing the probability of superior
performance within environments and regions. Exists only if a string is provided
for "reg"
\item \code{psp_env.mat} : a matrix containing the probability of superior
performance within environments. Exists only "reg" is NULL
\item \code{psp_env.plot} : a heatmap representing the probability of superior
performance of each genotype, within each environment
\item \code{psp_reg.plot} : a heatmap representing the probability of superior
performance of each genotype, within each region. Exists only if a string is
provided for "reg"
}
}
\description{
This function computates the  probabilities of superior performance within
environments and yields plots for better visualization.
All the plots are customizable, using ggplot2
}
\examples{
\dontrun{conds = marg_prob(data = df, trait = "GY", gen = "H",env = "L", extr_outs = outs,
                  reg = "Region", int = .2, save.df = TRUE, interactive = TRUE)}
}
