% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_prob.R
\name{cond_prob}
\alias{cond_prob}
\title{Conditional probabilities of superior performance}
\usage{
cond_prob(
  data,
  trait,
  gen,
  env,
  reg = NULL,
  extr_outs,
  int = 0.2,
  increase = TRUE,
  save.df = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the observations}

\item{trait}{A character representing the name of the column that
corresponds to the analysed variable}

\item{gen}{A character representing the name of the column that
corresponds to the evaluated genotypes}

\item{env}{A character representing the name of the column that
corresponds to the environments}

\item{reg}{A character representing the name of the column that
corresponds to the regions. NULL otherwise (default)}

\item{extr_outs}{An object from the \code{extr_outs} function}

\item{int}{A number representing the selection intensity
(superior limit = 1)}

\item{increase}{Logical: TRUE if genotypes with higher trait values are better.
FALSE otherwise.}

\item{save.df}{A logical value indicating if the data frames with the marginal
probability of each genotype and the pairwise probabilities should be saved at
the work directory.}

\item{interactive}{A logical value indicating if the plots should be interactive.
If TRUE, the function loads the \code{plotly} package and uses the \code{\link[plotly:ggplotly]{plotly::ggplotly()}}
command.}
}
\value{
The function returns a list with:
\itemize{
\item \code{conds_prob} : a matrix containing the probability of superior
performance within environments (and regions, if available).
\item \code{psp_env.plot} : a heatmap representing the probability of superior
performance of each genotype, within each environment
\item \code{psp_reg.plot} : a heatmap representing the probability of superior
performance of each genotype, within each region. Exists only if a string is
provided for "reg"
}
}
\description{
This function estimates the  probabilities of superior performance within
environments and yields plots for better visualization.
All the plots are customizable, using \code{ggplot2}
}
\details{
Using \code{cond_prob()}, you can estimate the probability of a genotype being amongst the
selected, based on a given selection intensity in a specific environment or
breeding region. If we let \eqn{\Omega_k} represent the subset of superior
genotypes in the \if{html}{\out{<i>}}kth\if{html}{\out{</i>}} environment, then the probability of the \if{html}{\out{<i>}}jth\if{html}{\out{</i>}}
genotype belonging to \eqn{\Omega_k} is:

\deqn{Pr(g_{jk} \in \Omega_k \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_{jk}^{(s)} \in \Omega \vert y)}}

where \eqn{S} is the total number of samples, and \eqn{I(g_{jk}^{(s)} \in \Omega \vert y)}
is an indicator variable mapping success (1) if \eqn{g_{jk}^{(s)}} exists in \eqn{\Omega},
and failure (0) otherwise.
}
\examples{
\dontrun{
mod = bayes_met(data = maize, gen = c("Hybrid", "normal", "cauchy"),
                env = c("Location", "normal", "cauchy"),
                rept = list(c("Rep", "normal", "cauchy"), c("Block", "normal", "cauchy")),
                trait = "GY", hyperparam = "default", sigma.dist = c("cauchy", "cauchy"),
                mu.dist = c("normal", "normal"), gei.dist = c("normal", "cauchy"),
                reg = list(c("Region", "normal", "cauchy"), c("normal", "cauchy")),
                res.het = T,
                iter = 100, cores = 2, chain = 2)
                #Reiterating: increase the number of iterations, cores and chains

outs = extr_outs(data = maize, trait = "GY", gen = "Hybrid", model = mod,
                 effects = c("r", "b", "l", "m", "g", "gl", "gm"),
                 nenv = 16, res.het = TRUE, check.stan.diag = TRUE)

conds = cond_prob(data = maize, trait = "GY", gen = "Hybrid", env = "Location",
                  extr_outs = outs, reg = "Region", int = .2,
                  increase = TRUE, save.df = TRUE, interactive = TRUE)
                  }
}
\references{
Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M.,
Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts
for cultivar recommendation in multi-environment trials. \if{html}{\out{<i>}}Theoretical and
Applied Genetics\if{html}{\out{</i>}}, 133(2):443-455. https://doi.org/10.1007/s00122-022-04041-y
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
