% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_outs.R
\name{extr_outs}
\alias{extr_outs}
\title{Extracting outputs of "rstan" models}
\usage{
extr_outs(
  data,
  trait,
  gen,
  model,
  effects,
  nenv,
  res.het = FALSE,
  check.stan.diag = FALSE,
  ...
)
}
\arguments{
\item{data}{A dataframe containing the observations}

\item{trait}{A character representing the name of the column that
corresponds to the analysed variable}

\item{gen}{A character representing the name of the column that
corresponds to the evaluated genotypes}

\item{model}{An object containing the Bayesian model fitted using rstan}

\item{effects}{A string vector containing the codes of the effects included
in the model. The codes are:
\itemize{
\item \code{r} : replicate effect
\item \code{b} : block inside replicate effect (if the trials are lattices)
\item \code{l} : environment (or location) effect
\item \code{m} : region effect (if it is available)
\item \code{g} : genetic effect
\item \code{gl} : genotype-by-location effect
\item \code{gm} : genotype-by-region effect
}}

\item{nenv}{The number of environments in the analysis}

\item{res.het}{Are the residual heterogeneous? Default is FALSE}

\item{check.stan.diag}{A logical value indicating whether the function should
extract some diagnostic using native stan functions.}

\item{...}{Passed to stan_diag (for more information, see stan_diag manual)}
}
\value{
The function returns a list with:
\itemize{
\item \code{post} : a list with the posterior of the effects, and the data
generated by the model
\item \code{variances} : a dataframe with the variances of each effect
\item \code{map} : a list with the maximum posterior values of each effect
\item \code{ppcheck} : a matrix containing the p-values of maximum, minimum,
median, mean and standard deviation; effective number of parameters, WAIC2
value and Rhat
\item \code{den_plot} a density plot comparing the generated data and the
real data. Note that this plot is customizable using the "ggplot2" library
}
}
\description{
This function extracts outputs of the Bayesian model fitted using "rstan",
and provides some diagnostics about the model
}
\examples{
\dontrun{
mod = bayes_met(data = maize, gen = c("Hybrid", "normal", "cauchy"),
                env = c("Location", "normal", "cauchy"),
                rept = list(c("Rep", "normal", "cauchy"), c("Block", "normal", "cauchy")),
                trait = "GY", hyperparam = "default", sigma.dist = c("cauchy", "cauchy"),
                mu.dist = c("normal", "normal"), gli.dist = c("normal", "cauchy"),
                reg = list(c("Region", "normal", "cauchy"), c("normal", "cauchy")),
                iter = 100, cores = 2, chain = 2)
                #Do not forget to increase the number of iterations, cores and chains

outs = extr_outs(data = maize, trait = "GY", gen = "Hybrid", model = mod,
                 effects = c("r", "b", "l", "m", "g", "gl", "gm"),
                 nenv = 16, res.het = TRUE, check.stan.diag = TRUE)
                 }
}
