% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maize.R
\docType{data}
\name{maize}
\alias{maize}
\title{Maize data set}
\format{
\subsection{\code{maize}}{

A data frame with 1,153 rows and 6 colums:
\describe{
\item{Region}{Regions containing the evaluated environments: TB, TR, TA, MN, and NE}
\item{Location}{Evaluated environments: E1 to E16}
\item{Rep}{Replicates: 1 and 2}
\item{Block}{Blocks within replicates: 1 to 12}
\item{Hybrid}{Evaluated hybrids: G1 to G36}
\item{GY}{Grain yield (ton)}
}
}
}
\source{
\describe{
Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M.,
Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts
for cultivar recommendation in multi-environment trials. \if{html}{\out{<i>}}Theoretical and
Applied Genetics\if{html}{\out{</i>}}, 133(2):443-455. https://doi.org/10.1007/s00122-022-04041-y
}
}
\usage{
maize
}
\description{
The data used by Dias et al. (2022) in the seminal paper (see reference). The
dataset comprised 32 single-cross hybrid plus four commercial checks (36 hybrids)
from late-stage breeding of Embrapa Maize and Sorghum evaluated in 2013 for grain
yield in 16 environments. These environments represent five tropical breeding
regions in Brazil.
}
\keyword{datasets}
