% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marg_prob.R
\name{marg_prob}
\alias{marg_prob}
\title{Marginal probabilities of superior performance}
\usage{
marg_prob(
  data,
  trait,
  gen,
  env,
  extr_outs,
  int = 0.2,
  increase = T,
  save.df = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{A dataframe containing the observations}

\item{trait}{A character representing the name of the column that
corresponds to the analysed variable}

\item{gen}{A character representing the name of the column that
corresponds to the evaluated genotypes}

\item{env}{A character representing the name of the column that
corresponds to the environments}

\item{extr_outs}{An object from the \code{extr_outs} function}

\item{int}{An integer representing the selection intensity
(superior limit = 1)}

\item{increase}{Logical: TRUE if genotypes with higher trait values are better.
FALSE otherwise.}

\item{save.df}{A logical value indicating if the data frames with the marginal
probability of each genotype and the pairwise probabilities should be saved at
the work directory.}

\item{interactive}{A logical value indicating if the plots should be interactive.
If TRUE, the function loads the \code{plotly} package and uses the \link[plotly:ggplotly]{plotly::ggplotly}
command.}
}
\value{
The function returns a list with:
\itemize{
\item \code{g_hpd} : a catterpillar plot representing the performance of
each genotype, and their respective confidence interval (95\% (thick) and 97.5\% (thin))
\item \code{marg_prob.df} : a dataframe with the marginal probability of superior
performance of each genotype
\item \code{marg_prob.plot} : a bar plot with the marginal probability of superior
performance of each genotype
\item \code{pair_prob.df} : a dataframe with the pairwise marginal probability
of superior performance between genotypes
\item \code{pair_prob.plot} a heatmap representing the pairwise probability of
superior performance
}
}
\description{
This function estimates the marginal probabilities of superior performance and
yields plots for better visualization. All the plots are customizable, using \code{ggplot2}
}
\details{
Using \code{marg_prob()}, you can estimate the probability of a genotype being amongst the
selected, based on a given selection intensity. If we let \eqn{\Omega} represent the
subset of superior genotypes, then the probability of the \if{html}{\out{<i>}}jth\if{html}{\out{</i>}} genotype belonging to
\eqn{\Omega} is:

\deqn{Pr(g_j \in \Omega \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} \in \Omega \vert y)}}

where \eqn{S} is the total number of samples, and \eqn{I(g_j^{(s)} \in \Omega \vert y)}
is an indicator variable mapping success (1) if \eqn{g_j^{(s)}} exists in \eqn{\Omega},
and failure (0) otherwise.

This function also estimates the pairwise probability of superior performance, i.e.
the probability of a given genotype \if{html}{\out{<i>}}j\if{html}{\out{</i>}} being superior to genotype \if{html}{\out{<i>}}i\if{html}{\out{</i>}}:

\deqn{Pr(g_j > g_i \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} > g_i^{(s)} \vert y)}}
}
\examples{
\dontrun{
mod = bayes_met(data = maize, gen = c("Hybrid", "normal", "cauchy"),
                env = c("Location", "normal", "cauchy"),
                rept = list(c("Rep", "normal", "cauchy"), c("Block", "normal", "cauchy")),
                trait = "GY", hyperparam = "default", sigma.dist = c("cauchy", "cauchy"),
                mu.dist = c("normal", "normal"), gli.dist = c("normal", "cauchy"),
                reg = list(c("Region", "normal", "cauchy"), c("normal", "cauchy")),
                res.het = T,
                iter = 100, cores = 2, chain = 2)
                #Remember, increase the number of iterations, cores and chains

outs = extr_outs(data = maize, trait = "GY", gen = "Hybrid", model = mod,
                 effects = c("r", "b", "l", "m", "g", "gl", "gm"),
                 nenv = 16, res.het = T, check.stan.diag = TRUE)

margs = marg_prob(data = maize, trait = "GY", gen = "Hybrid",env = "Location",
                  increase = TRUE,extr_outs = outs, int = .2,
                  save.df = TRUE, interactive = TRUE)
                  }

}
\references{
Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M.,
Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts
for cultivar recommendation in multi-environment trials. \if{html}{\out{<i>}}Theoretical and
Applied Genetics\if{html}{\out{</i>}}, 133(2):443-455. https://doi.org/10.1007/s00122-022-04041-y
}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
