% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpsi.R
\name{print.bpsi}
\alias{print.bpsi}
\title{Print an object of class \code{bpsi}}
\usage{
\method{print}{bpsi}(x)
}
\arguments{
\item{x}{An object of class \code{bpsi}}
}
\description{
Print a \code{bpsi} object in R console
}
\seealso{
\code{\link[=bpsi]{bpsi()}}
}
\author{
José Tiago Barroso Chagas
}
