% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_outs.R
\name{print.extr}
\alias{print.extr}
\title{Print an object of class \code{extr}}
\usage{
\method{print}{extr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{extr}}

\item{...}{currently not used}
}
\description{
Print a \code{extr} object in R console
}
\seealso{
\link{extr_outs}
}
