% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_sup.R
\name{prob_sup}
\alias{prob_sup}
\title{Probabilities of superior performance and stability}
\usage{
prob_sup(
  data,
  trait,
  gen,
  env,
  reg = NULL,
  extr_outs,
  int,
  increase = TRUE,
  save.df = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{A data frame containing the phenotypic data}

\item{trait, gen, env}{A string. The name of the columns that corresponds to
the variable, genotype and environment information, respectively}

\item{reg}{A string or NULL. If the data set has information about regions,
\code{reg} will be a string with the name of the column that corresponds to the
region information. Otherwise, \code{reg = NULL} (default).}

\item{extr_outs}{An object from the \code{\link[=extr_outs]{extr_outs()}} function}

\item{int}{A number representing the selection intensity
(between 0 and 1)}

\item{increase}{Logical. Indicates the direction of the selection.
\code{TRUE} (default) for increasing the trait value, \code{FALSE} otherwise.}

\item{save.df}{Logical. Should the data frames be saved in the work directory?
\code{TRUE} for saving, \code{FALSE} (default) otherwise.}

\item{interactive}{Logical. Should ggplots be converted into interactive plots?
If \code{TRUE}, the function loads the \code{plotly} package and uses the \code{\link[plotly:ggplotly]{plotly::ggplotly()}}
command.}
}
\value{
The function returns two lists, one with the \code{marginal} probabilities, and
another with the \code{conditional} probabilities.

The \code{marginal} list has:
\itemize{
\item \code{df} : A list of data frames containing the calculated probabilities:
\itemize{
\item \code{perfo}: the probabilities of superior performance
\item \code{pair_perfo}: the pairwise probabilities of superior performance
\item \code{stabi}: the probabilities of superior stability. When \code{reg} is not
\code{NULL}, \code{stabi} is divided into \code{stabi_gl} for the stability across environments, and
\code{stabi_gm} for the stability across regions
\item \code{pair_stabi}: the pairwise probabilities of superior stability, which
is also divided into \code{pair_stabi_gl} and \code{pair_stabi_gm} when \code{reg} is not \code{NULL}
\item \code{joint_prob}: the joint probabilities of superior performance and stability
}
\item \code{plot} : A list of ggplots illustrating the outputs:
\itemize{
\item \code{g_hpd}: a caterpillar plot representing the marginal genetic value of
each genotype, and their respective high posterior density (95\% represented by the
thick line, and 97.5\% represented by the thin line)
\item \code{perfo}: a bar plot illustrating the probabilities of superior performance
\item \code{pair_perfo}: a heatmap representing the pairwise probability of superior
performance (the probability of genotypes at the \emph{x}-axis being superior
to those on \emph{y}-axis)
\item \code{stabi}: a bar plot with the probabilities of superior stability. Like the data frames,
when \code{reg} is not \code{NULL}, two different plots are generated, one for the stability across
environments (\code{stabi_gl}), and another for the stability across regions (\code{stabi_gm})
\item \code{pair_stabi}: a heatmap with the pairwise probabilities of superior stability
(also divided into two different plots when \code{reg} is not \code{NULL}). This plot represents
the probability of genotypes at the \emph{x}-axis being superior
to those on \emph{y}-axis
\item \code{joint_prob}: a plot with the probabilities of superior performance,
probabilities of superior stability and the joint probabilities of superior
performance and stability.
}
}

The \code{conditional} list has:
\itemize{
\item \code{df} : A list with:
\itemize{
\item \code{perfo}: a data frame containing the probabilities of superior performance
within environments. It also has the probabilities of superior performance with regions
if \code{reg} is not \code{NULL}.
\item \code{pair_perfo}: a list with the pairwise probabilities of superior performance
within environments. If \code{reg} is not \code{NULL}, two lists are generated.
}
\item \code{plot} : A list with:
\itemize{
\item \code{perfo}: a heatmap with the probabilities of superior performance within
environments. If \code{reg} is not \code{NULL}, there will be two heatmaps: \code{perfo_env}
for the probabilities of superior performance within environments, and \code{perfo_reg}
for the same probabilities within regions.
\item \code{pair_perfo}: a list of heatmaps representing the pairwise probability of superior
performance within environments. If \code{reg} is not \code{NULL}, there will be two lists: \code{pair_perfo_env}
for the probabilities of superior performance within environments, and \code{pair_perfo_reg}
for the same probabilities within regions. The interpretation is the same as in the
\code{pair_perfo} in the \code{marginal} list: the probability of genotypes at the \emph{x}-axis being superior
to those on \emph{y}-axis.
}
}
}
\description{
This function estimates the probabilities of superior performance and stability
across environments (\code{marginal} output). It also computes the probabilities
of superior performance within environments (\code{conditional} output).
}
\details{
Probabilities provide the risk of recommending a candidate for a target
population of environments or for a specific environment. The function \code{prob_sup()}
computes the probabilities of superior performance (agronomic stability,
or previsibility), and the probabilities of superior stability (ecological
stability, or invariance):

\itemize{\item Probability of superior performance}

Let \eqn{\Omega} represent the subset of selected genotypes based on their
performance across environments. A given genotype \eqn{j} will belong to \eqn{\Omega}
if its genetic marginal value (\eqn{g_j}) is among the top \eqn{V}, with \eqn{V}
being the selection intensity. \code{prob_sup()} leverages the Monte Carlo discretized sampling
from the posterior distribution to emulate the occurrence of \eqn{S} trials. Thence,
the probability of the \eqn{j^{th}} genotype belonging to \eqn{\Omega} is the
ratio of success (\eqn{g_j \in \Omega}) and the total number of sampled events
(the emulated \eqn{S} trials):

\deqn{Pr(g_j \in \Omega \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} \in \Omega \vert y)}}

where \eqn{S} is the total number of samples (\eqn{s = 1, 2, ..., S}),
and \eqn{I(g_j^{(s)} \in \Omega \vert y)} is an indicator variable mapping
success (1) if \eqn{g_j^{(s)}} exists in \eqn{\Omega} in the \eqn{s^{th}} sample,
and failure (0) otherwise. \eqn{S} is conditioned to the number of iterations and chains
previously set at \code{\link[=bayes_met]{bayes_met()}}.

The same idea can be applied for each environment: let \eqn{\Omega_k} represent
the subset of superior genotypes in the \eqn{k^{th}} environment, then the
probability of the \eqn{j^{th}} genotype belonging to \eqn{\Omega_k} is

\deqn{Pr(g_{jk} \in \Omega_k \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_{jk}^{(s)} \in \Omega_k \vert y)}}

where \eqn{I(g_{jk}^{(s)} \in \Omega_k \vert y)} is an indicator variable mapping
success (1) if \eqn{g_{jk}^{(s)}} exists in \eqn{\Omega_k}, and failure (0) otherwise.
Note that \eqn{g_{jk} = g_j + ge_{jk}}, i.e., the genetic marginal value is summed to
the genetic value of \eqn{j} when planted specifically at environment \eqn{k}
(genotype-by-environment interaction effect).

Pairwise comparisons are also available in \code{prob_sup()}, both across and within environments.
Leveraging the same samples, \code{prob_sup()} estimates the probability of a genotype
\eqn{j} performing better than a genotype \eqn{i}. Across environments, the
equation is the following:

\deqn{Pr(g_j > g_i \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} > g_i^{(s)} \vert y)}}

or

\deqn{Pr(g_j < g_i \vert y) = \frac{1}{S}\sum_{s=1}^S{I(g_j^{(s)} < g_i^{(s)} \vert y)}}

The first equation is applied when the aim is to increase the trait value (\code{increase = T}).
Conversely, the second equation is used when \code{increase = F}. Within environments,
\eqn{g_j} is switched by \eqn{g_{jk}}.

\itemize{\item Probability of superior stability}

The equations described above are related to the genotypes' previsibility, an
agronomic way of looking at stability. Another point of view is to consider the
invariance as a measure of stability. Making a direct analogy with Shukla's (1972)
theory, a stable genotype would be the one that had the lower variance of the
genotype-by-environment interaction effect, \eqn{var(ge_{jk})}. Using the same
ideas as above, the probability of superior stability is given by:

\deqn{Pr [var(ge_{jk})\in \Omega \vert y] = \frac{1}{S} \sum_{s=1}^S{I [var(ge_{jk}^{(s)})\in \Omega \vert y]} }

where \eqn{I[var(ge_{jk}^{(s)})\in \Omega \vert y]} is an indicator variable
mappings success (1) or failure (0).

\code{prob_sup()} also provides means to directly compare the stability of two genotypes.
The pairwise probability of superior stability is given by:

\deqn{Pr[var(ge_{jk}) < var(ge_{ik}) \vert y] = \frac{1}{S} \sum_{s=1}^S{I [var(ge_{jk}^{(s)}) < var(ge_{ik}^{(s)}) \vert y]} }

Note that \eqn{j} will be superior to \eqn{i} if it has a \strong{lower}
variance of the genotype-by-environment interaction effect. This is true regardless
if \code{increase} is set to \code{TRUE} or \code{FALSE}.

Finally, by assuming independence between the genetic main effect and the
genotype-by-environment effect variance, \code{prob_sup} computes the joint probability
of superior performance and stability:

\deqn{Pr[g_j \in \Omega, var(ge_{jk} \in \Omega)] = Pr(g_j \in \Omega \vert y) \times Pr[var(ge_{jk}) \in \Omega \vert y]}

The estimation of these probabilities are strictly related to some key question that
constantly arises in plant breeding:
\itemize{
\item \strong{What is the risk of recommending a selection candidate for a target population of environments?}
\item \strong{What is the probability of a given selection candidate having good performance if
recommended to a target population of environments? And for a specific environment?}
\item \strong{What is the probability of a given selection candidate having better performance
than a cultivar check in the target population of environments? And in specific environments?}
\item \strong{How probable is it that a given selection candidate performs similarly across environments?}
\item \strong{What are the chances that a given selection candidate is more stable
than a cultivar check in the target population of environments?}
\item \strong{What is the probability that a given selection candidate having a
superior and invariable performance across environments?}
}
}
\examples{
\dontrun{
mod = bayes_met(data = soy,
                gen = "Gen",
                env = "Env",
                repl = NULL,
                reg = "Reg",
                res.het = F,
                trait = "Y",
                iter = 2000, cores = 1, chains = 4)

outs = extr_outs(data = soy, trait = "Y", gen = "Gen", model = mod,
                 effects = c('l','g','gl','m','gm'),
                 nenv = length(unique(soy$Env)),
                 probs = c(0.05, 0.95), check.stan.diag = TRUE)

results = prob_sup(data = soy, trait = "Y", gen = "Gen", env = "Env",
                   extr_outs = outs, reg = 'Reg', int = .2,
                   increase = T, save.df = T, interactive = T)
}

}
\references{
Dias, K. O. G, Santos J. P. R., Krause, M. D., Piepho H. -P., Guimarães, L. J. M.,
Pastina, M. M., and Garcia, A. A. F. (2022). Leveraging probability concepts
for cultivar recommendation in multi-environment trials. \emph{Theoretical and
Applied Genetics}, 133(2):443-455. \doi{10.1007/s00122-022-04041-y}

Shukla, G. K. (1972) Some statistical aspects of partioning genotype environmental
componentes of variability. \emph{Heredity}, 29:237-245. \doi{10.1038/hdy.1972.87}
}
