% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProbBreed-package.R
\docType{package}
\name{ProbBreed-package}
\alias{ProbBreed}
\alias{ProbBreed-package}
\title{ProbBreed: Probability Theory for Selecting Candidates in Plant Breeding}
\description{
Use probability theory under the Bayesian framework for calculating the risk of selecting candidates in a multi-environment context [Dias et al. (2022) \doi{10.1007/s00122-022-04041-y}]. Contained are functions used to fit a Bayesian multi-environment model (based on the available presets), extract posterior values and maximum posterior values, compute the variance components, check the model’s convergence, and calculate the probabilities. For both across and within-environments scopes, the package computes the probability of superior performance and the pairwise probability of superior performance. Furthermore, the probability of superior stability and the pairwise probability of superior stability across environments is estimated. A joint probability of superior performance and stability is also provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saulo-chaves/ProbBreed}
  \item \url{https://saulo-chaves.github.io/ProbBreed_site/}
  \item \url{https://saulo-chaves.github.io/ProbBreed/}
  \item Report bugs at \url{https://github.com/saulo-chaves/ProbBreed/issues}
}

}
\author{
\strong{Maintainer}: Saulo Chaves \email{saulo.chaves@ufv.br} (\href{https://orcid.org/0000-0002-0694-1798}{ORCID})

Authors:
\itemize{
  \item Kaio Dias \email{kaio.o.dias@ufv.br} (\href{https://orcid.org/0000-0002-9171-1021}{ORCID}) [copyright holder]
  \item Matheus Krause \email{mdkrause@iastate.edu} (\href{https://orcid.org/0000-0003-2411-9287}{ORCID})
}

}
\keyword{internal}
